Add optional to disable strip path and enable uppercasing the name
This commit adds two new cli flags:
1. -strip-path Will remove the AWS_ENV_PATH from the environment
variable, useful when you need to call aws-env multiple times and want
to ensure that the environment variables exposed won't conflict with
each other.
2. -uppercase Changes the capitalisation of the variable's name to
all-caps. It is common for ssm parameters to be in all-lowercase whereas
we obviously want the environment variables to be in uppercase.